move protobuf aliases to lib.rs, update imports
