+++ rust/tw_bitcoin/src/entry.rs
@@ -15,20 +15,6 @@ use tw_misc::traits::ToBytesVec;
 use tw_proto::BitcoinV2::Proto;
 use tw_proto::Utxo::Proto as UtxoProto;
 
-// Convenience aliases.
-pub use aliases::*;
-pub mod aliases {
-    use super::Proto;
-
-    pub type ProtoOutputRecipient<'a> = Proto::mod_Output::OneOfto_recipient<'a>;
-    pub type ProtoOutputBuilder<'a> = Proto::mod_Output::mod_OutputBuilder::OneOfvariant<'a>;
-    pub type ProtoPubkeyOrHash<'a> = Proto::mod_ToPublicKeyOrHash::OneOfto_address<'a>;
-    pub type ProtoRedeemScriptOrHash<'a> =
-        Proto::mod_Output::mod_OutputRedeemScriptOrHash::OneOfvariant<'a>;
-    pub type ProtoInputRecipient<'a> = Proto::mod_Input::OneOfto_recipient<'a>;
-    pub type ProtoInputBuilder<'a> = Proto::mod_Input::mod_InputBuilder::OneOfvariant<'a>;
-}
-
 pub struct Address(pub bitcoin::address::Address<NetworkChecked>);
 
 impl Display for Address {
